+++ Paradigm/src/utopia/paradigm/motion/template/AccelerationLike.scala
@@ -5,7 +5,7 @@ import utopia.flow.time.TimeExtensions._
 import utopia.paradigm.enumeration.Axis
 import utopia.paradigm.motion.motion1d.LinearAcceleration
 import utopia.paradigm.shape.template.HasDimensions.HasDoubleDimensions
-import utopia.paradigm.shape.template.{Dimensional, DoubleVectorLike, VectorProjectable2}
+import utopia.paradigm.shape.template.{Dimensional, DoubleVectorLike, VectorProjectable}
 
 import scala.concurrent.duration.Duration
 
@@ -19,7 +19,7 @@ import scala.concurrent.duration.Duration
   */
 trait AccelerationLike[X <: DoubleVectorLike[X], V <: VelocityLike[X, V], +Repr <: Change[V, _]]
 	extends Change[V, Repr] with Combinable[Change[V, _], Repr] with Dimensional[LinearAcceleration, Repr]
-		with VectorProjectable2[Repr] with CanBeAboutZero[Change[Change[HasDoubleDimensions, _], _], Repr]
+		with VectorProjectable[Repr] with CanBeAboutZero[Change[Change[HasDoubleDimensions, _], _], Repr]
 {
 	// ABSTRACT	-----------------------