+++ Paradigm/src/utopia/paradigm/motion/template/VelocityLike.scala
@@ -4,7 +4,7 @@ import utopia.flow.operator.{CanBeAboutZero, Combinable, LinearScalable}
 import utopia.flow.time.TimeExtensions._
 import utopia.paradigm.motion.motion1d.{LinearAcceleration, LinearVelocity}
 import utopia.paradigm.shape.template.HasDimensions.HasDoubleDimensions
-import utopia.paradigm.shape.template.{Dimensional, DoubleVectorLike, VectorProjectable2}
+import utopia.paradigm.shape.template.{Dimensional, DoubleVectorLike, VectorProjectable}
 
 import scala.concurrent.duration.Duration
 
@@ -18,7 +18,7 @@ import scala.concurrent.duration.Duration
 trait VelocityLike[X <: DoubleVectorLike[X], +Repr <: Change[X, Repr]]
 	extends Change[X, Repr] with LinearScalable[Repr] with Combinable[Change[HasDoubleDimensions, _], Repr]
 		with CanBeAboutZero[Change[HasDoubleDimensions, _], Repr] with Dimensional[LinearVelocity, Repr]
-		with VectorProjectable2[Repr]
+		with VectorProjectable[Repr]
 {
 	// ABSTRACT	-----------------