+++ Scribe/Scribe-Api/src/utopia/scribe/api/database/access/single/logging/issue/DbIssue.scala
@@ -102,8 +102,7 @@ object DbIssue extends SingleRowModelAccess[Issue] with UnconditionalView with I
 			timeRange.start)
 		val variant = (variantDependenciesType match {
 			// Case: There is a chance that the variant already exists => Checks for duplicates before inserting
-			case Last =>
-				DbIssueVariant.findMatching(variantData).toRight { variantModel.insert(variantData) }
+			case Last => DbIssueVariant.findMatching(variantData).toRight { variantModel.insert(variantData) }
 			// Case: It's impossible that the variant would already exist => Inserts a new variant
 			case First => Left(variantModel.insert(variantData))
 		}).either