+++ Firmament/Changes.md
@@ -39,6 +39,7 @@
 - **Window** mouse move events are now only fired while the window has focus, except in windows that can't gain focus.
 - **SingleColorIconCache** now logs errors and appends ".png" to file names in case the file extension is not specified
 - **SingleColorIcon** now extends **MayBeEmpty**
+- **CachingStackable** now rounds stack size values before caching them (this update may be removed in the future)
 
 ## v1.0 - 01.05.2023
 Initial version