+++ Paradigm/Changes.md
@@ -1,9 +1,12 @@
 # Utopia Paradigm - List of Changes
 
-## v1.3.1 (in development)
+## v1.4 (in development)
+### Breaking Changes
+- Renamed **VectorProjectable2** to **VectorProjectable** (leftover from an earlier update)
 ### New Features
 - Added **Adjustment** and **SizeAdjustable** for more convenient size change functions
 - Added **FromDirectionFactory** and **FromAlignmentFactory** -traits
+- Added **RoundingDouble** wrapper class in preparation of rounding shapes (which have not yet been implemented)
 ### New Methods
 - **Dimensions**
   - Added `.mapWithZero(...)(...)`